Temp commit for PC switch

+++ Calculator/src/main/java/com/lorgen/calculator/numerical/NumberUtils.java
@@ -1,7 +1,7 @@
 package com.lorgen.calculator.numerical;
 
 public class NumberUtils {
-    public static boolean isPrime(int number) {
+    public static boolean isPrime(long number) {
         if (number % 2 == 0) return false;
         for (int i = 3; i * i <= number; i += 2) if (number % i == 0) return false;
         return true;
